feat: 데미지 입히기 추가. 치명타 보정 추가.
- 테스트 필요.
