+++ ProjectP/Assets/02.Scripts/NewWeapon/Projectiles/RifleProjectile.cs.meta
@@ -1,3 +1,3 @@
 fileFormatVersion: 2
-guid: 065de498c1ce44df87d0ba0d82e383aa
-timeCreated: 1774888011
+guid: c172cd2ac7bb47ecbea4ab0d1c371f16
+timeCreated: 1774888056
