--- ProjectP/Assets/02.Scripts/NewWeapon/Projectiles/StraightMovement.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: c172cd2ac7bb47ecbea4ab0d1c371f16
-timeCreated: 1774888056
